requirements.txt
a2wsgi==1.10.10
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.12.0
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-mongo==3.2.0
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.1.0
Babel==2.13.0
backoff==1.10.0
blinker==1.6.3
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron_descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.4
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
fastapi==0.119.1
fastapi-cli==0.0.14
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.0
fsspec==2025.9.0
google-re2==1.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenback==1.2.1
greenlet==3.2.4
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httptools==0.7.1
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
libcst==1.8.5
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.0.4
natsort==8.4.0
numpy==2.3.4
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.3.3
pathlib_abc==0.5.2
pathspec==0.11.2
pendulum==2.1.2
pluggy==1.3.0
prison==0.2.1
protobuf==4.21.12
psutil==5.9.5
psycopg2-binary==2.9.11
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pymongo==4.15.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==6.4.0
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.3.0
rpds-py==0.10.4
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
spotipy==2.25.1
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
typer==0.20.0
types-requests==2.32.4.20250913
typing_extensions==4.8.0
tzdata==2025.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal_pathlib==0.3.4
urllib3==1.26.17
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0

# Streamlit 核心
streamlit==1.28.1

# 圖表和視覺化
plotly==5.17.0
pandas==2.1.3